Refactored Facebook into it's own folder.
