+++ src/components/pages/user/Login.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { Button, FormGroup, Input, Label } from 'reactstrap';
 import { Auth } from 'aws-amplify';
 import 'styles/pages/user/login.scss';
-import { loadFacebookSDK } from '../../utils/Facebook';
-import FacebookButton from '../../utils/FacebookButton';
+import { loadFacebookSDK } from '../../utils/Facebook/FacebookSDK';
+import FacebookButton from '../../utils/Facebook/FacebookButton';
 
 interface Props {
   history: any; // tslint:disable-line: no-any
